fixed duplicated test file
