+++ src/game/state.spec.ts
@@ -1,7 +1,5 @@
-import { Board, getLastPlayedTile } from "../board/board";
-import { getPlayableTilesForNextMove } from "../move/move";
-import { initGameState, winGame, checkIfGameWon } from "./state";
-
+import { Board } from "../board/board";
+import { initGameState, winGame, checkIfGameWon, checkIfDraw } from "./state";
 const mockFilledBoard: Board = [
   [
     undefined,
@@ -96,5 +94,26 @@ describe("checkIfGameIsWon", () => {
     let boardWithPossibleMove = JSON.parse(JSON.stringify(mockFilledBoard));
     boardWithPossibleMove[0][4] = { symbol: "D", color: "blue" };
     expect(checkIfGameWon(initGameState(), boardWithPossibleMove)).toBeFalsy();
+  });
+});
+
+describe("checkIfDraw", () => {
+  it("should be false if there is a winner", () => {
+    let gameState = initGameState();
+    gameState.turnNumber = 36;
+    gameState.winner = "black";
+    expect(checkIfDraw(gameState)).toBeFalsy();
+  });
+
+  it("should be true for turn number after 35", () => {
+    let gameState = initGameState();
+    gameState.turnNumber = 36;
+    expect(checkIfDraw(gameState)).toBeTruthy();
+  });
+
+  it("should be false for turn number 35 and below", () => {
+    let gameState = initGameState();
+    gameState.turnNumber = 35;
+    expect(checkIfDraw(gameState)).toBeFalsy();
   });
 });
